feat(config): changes for exporting backend
