+++ backend/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "lumetv-backend",
+  "version": "1.0.0",
+  "type": "module",
+  "main": "index.ts",
+  "scripts": {
+    "start": "node --loader ts-node/esm index.ts"
+  },
+  "dependencies": {
+    "tmi.js": "^1.8.5",
+    "ws": "^8.13.0"
+  }
+}
