+++ backend/tsconfig.json
@@ -1,0 +1,102 @@
+{
+	"compilerOptions": {
+		"forceConsistentCasingInFileNames": true,
+		"strict": true,
+		"noEmit": true,
+		"target": "ESNext",
+		"module": "ESNext",
+		"moduleResolution": "Bundler",
+		"allowJs": true,
+		"resolveJsonModule": true,
+		"jsx": "preserve",
+		"allowSyntheticDefaultImports": true,
+		"jsxFactory": "h",
+		"jsxFragmentFactory": "Fragment",
+		"paths": {
+			"#imports": ["./types/nitro-imports"],
+			"~/*": ["../*"],
+			"@/*": ["../*"],
+			"~~/*": ["../*"],
+			"@@/*": ["../*"],
+			"nitropack/types": ["../node_modules/nitropack/types"],
+			"nitropack/runtime": ["../node_modules/nitropack/runtime"],
+			"nitropack": ["../node_modules/nitropack"],
+			"defu": ["../node_modules/defu"],
+			"h3": ["../node_modules/h3"],
+			"consola": ["../node_modules/consola"],
+			"ofetch": ["../node_modules/ofetch"],
+			"@nuxt/devtools": ["../node_modules/@nuxt/devtools"],
+			"@vue/runtime-core": ["../node_modules/@vue/runtime-core"],
+			"@vue/compiler-sfc": ["../node_modules/@vue/compiler-sfc"],
+			"unplugin-vue-router/client": ["../node_modules/unplugin-vue-router/client"],
+			"@nuxt/schema": ["../node_modules/@nuxt/schema"],
+			"nuxt": ["../node_modules/nuxt"],
+			"vite/client": ["../node_modules/vite/client"],
+			"#shared": ["../shared"],
+			"assets": ["../assets"],
+			"assets/*": ["../assets/*"],
+			"public": ["../public"],
+			"public/*": ["../public/*"],
+			"#build": ["./"],
+			"#build/*": ["./*"],
+			"#internal/nuxt/paths": ["../node_modules/nuxt/dist/core/runtime/nitro/utils/paths"],
+			"#content/components": ["./content/components"],
+			"#content/manifest": ["./content/manifest"],
+			"#mdc-configs": ["./mdc-configs"],
+			"#mdc-highlighter": ["./mdc-highlighter"],
+			"#mdc-imports": ["./mdc-imports"],
+			"#image": ["../node_modules/@nuxt/image/dist/runtime"],
+			"#image/*": ["../node_modules/@nuxt/image/dist/runtime/*"],
+			"#nuxt-scripts-validator": ["../node_modules/@nuxt/scripts/dist/runtime/validation/valibot"],
+			"#nuxt-scripts": ["../node_modules/@nuxt/scripts/dist/runtime"],
+			"#nuxt-scripts/*": ["../node_modules/@nuxt/scripts/dist/runtime/*"],
+			"#ui": ["../node_modules/@nuxt/ui/dist/runtime"],
+			"#ui/*": ["../node_modules/@nuxt/ui/dist/runtime/*"],
+			"#color-mode-options": ["./color-mode-options.mjs"],
+			"#unhead/composables": ["../node_modules/nuxt/dist/head/runtime/composables/v3"],
+			"#nuxt-icon-server-bundle": ["./nuxt-icon-server-bundle"],
+			"#content/dump": ["./content/database.compressed"],
+			"#content/adapter": ["./db0/connectors/better-sqlite3"],
+			"#content/local-adapter": ["./db0/connectors/better-sqlite3"]
+		},
+		"lib": ["esnext", "webworker", "dom.iterable"]
+	},
+	"include": [
+		"./content/types.d.ts",
+		"./types/nitro-nuxt.d.ts",
+		"../node_modules/@nuxtjs/mdc/runtime/server",
+		"../node_modules/@nuxt/content/runtime/server",
+		"../node_modules/@nuxt/eslint/runtime/server",
+		"../node_modules/@nuxt/fonts/runtime/server",
+		"../node_modules/@nuxt/icon/runtime/server",
+		"../node_modules/@nuxt/image/runtime/server",
+		"../node_modules/@nuxt/scripts/runtime/server",
+		"../node_modules/@nuxt/test-utils/runtime/server",
+		"../node_modules/@nuxtjs/color-mode/runtime/server",
+		"../node_modules/@nuxt/ui/runtime/server",
+		"../node_modules/@pinia/nuxt/runtime/server",
+		"../node_modules/@nuxt/devtools/runtime/server",
+		"../node_modules/@nuxt/telemetry/runtime/server",
+		"./types/nitro.d.ts",
+		"../**/*",
+		"../server/**/*"
+	],
+	"exclude": [
+		"../node_modules",
+		"../node_modules/nuxt/node_modules",
+		"../node_modules/@nuxtjs/mdc/node_modules",
+		"../node_modules/@nuxt/content/node_modules",
+		"../node_modules/@nuxt/eslint/node_modules",
+		"../node_modules/@nuxt/fonts/node_modules",
+		"../node_modules/@nuxt/icon/node_modules",
+		"../node_modules/@nuxt/image/node_modules",
+		"../node_modules/@nuxt/scripts/node_modules",
+		"../node_modules/@nuxt/test-utils/node_modules",
+		"../node_modules/@nuxtjs/color-mode/node_modules",
+		"../node_modules/@nuxt/ui/node_modules",
+		"../node_modules/@pinia/nuxt/node_modules",
+		"../node_modules/@nuxt/devtools/node_modules",
+		"../node_modules/@nuxt/telemetry/node_modules",
+		"../dist"
+	]
+}
