+++ components/CategoryCard.vue
